requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyYAML
psycopg2-binary
